requirements.txt
streamlit
python-dotenv
mcp
google-genai
